+++ packages/builder/src/mrl/index.ts
@@ -1,6 +1,4 @@
 export * from './MrlBuilder';
 export * from './MrlBuilder.interfaces';
 export * from './MrlBuilder.constants';
-export * from './wormhole/wormhole/wormholeFactory';
-export * from './wormhole/wormhole/WormholeConfig';
-export * from './wormhole/extrinsic/ethereumXcm/BatchContractAbi';
+export * from './providers/wormhole';
