+++ packages/builder/src/mrl/providers/wormhole/extrinsic/polkadotXcm/polkadotXcm.ts
@@ -4,9 +4,9 @@ import {
   AssetAmount,
   EvmParachain,
 } from '@moonbeam-network/xcm-types';
-import { ExtrinsicBuilder } from '../../../../extrinsic/ExtrinsicBuilder';
-import { ExtrinsicConfig } from '../../../../types/substrate/ExtrinsicConfig';
-import { MrlConfigBuilder } from '../../../MrlBuilder.interfaces';
+import { ExtrinsicBuilder } from '../../../../../extrinsic/ExtrinsicBuilder';
+import { ExtrinsicConfig } from '../../../../../types/substrate/ExtrinsicConfig';
+import { MrlConfigBuilder } from '../../../../MrlBuilder.interfaces';
 
 // TODO: Can we move them somewhere?
 const BUY_EXECUTION_FEE = 100_000_000_000_000_000n;
